requirements.txt
# Core LangChain and LangGraph# Core
langchain>=0.1.0
langgraph>=0.0.40
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-community>=0.0.20

# Browser Automation
playwright>=1.40.0
browser-use>=0.11.0

# Frontend
streamlit>=1.30.0

# Configuration
python-dotenv>=1.0.0

# Logging
structlog>=23.1.0

# Security
cryptography>=41.0.0

# Utilities
pillow>=10.0.0
pydantic>=2.5.0
typing-extensions>=4.9.0
